requirements.txt
aws-cdk-lib>=2.87.0
constructs==10.2.69
cdk-nag==2.27.82